+++ controls/models.py
@@ -12,6 +12,7 @@ from copy import deepcopy
 from django.db import transaction
 
 BASELINE_PATH = os.path.join(os.path.dirname(__file__),'data','baselines')
+ORGPARAM_PATH = os.path.join(os.path.dirname(__file__),'data','org_defined_parameters')
 
 class ImportRecord(models.Model):
     created = models.DateTimeField(auto_now_add=True, db_index=True)
@@ -615,6 +616,68 @@ class Baselines (object):
         else:
             print("Requested baseline name not found in baselines_key data file")
             return False
+
+    @property
+    def body(self):
+        return self.legacy_imp_smt
+
+class OrgParams (object):
+    """Represent list of organizational defined parameters. Temporary class to work with default org params."""
+    def __init__(self):
+        global ORGPARAM_PATH
+        self.file_path = ORGPARAM_PATH
+        self.odp_keys = self._list_keys()
+        # self.index = self._build_index()
+
+        # Usage
+            # from controls.models import OrgParams
+            # odp = OrgParams()
+            # odp.odp_keys
+            # odp53 = odp._load_json('org_params_mod_fedramp')
+            # odp53['ac-1_prm_2']
+            # # Returns 'at least every 3 years'
+            # odp.get_org_params('org_params_mod_fedramp')
+
+    def _list_files(self):
+        return [
+            'org_params_low_fedramp.json',
+            'org_params_mod_fedramp.json',
+            'org_params_high_fedramp.json',
+        ]
+
+    def _list_keys(self):
+        return [
+            'org_params_low_fedramp',
+            'org_params_mod_fedramp',
+            'org_params_high_fedramp'
+        ]
+
+    def _load_json(self, odp_key):
+        """Read baseline file - JSON"""
+        # TODO Escape odp_key
+        self.data_file = odp_key + ".json"
+        data_file = os.path.join(self.file_path, self.data_file)
+        # Does file exist?
+        if not os.path.isfile(data_file):
+            print("ERROR: {} does not exist".format(data_file))
+            return False
+        # Load file as json
+        try:
+            with open(data_file, 'r') as json_file:
+                data = json.load(json_file)
+            return data
+        except:
+            print("ERROR: {} could not be read or could not be read as json".format(data_file))
+            return False
+
+    def get_org_params(self, odp_key):
+        """Return org defined param ids odp_key"""
+        if odp_key in self.odp_keys:
+            data = self._load_json(odp_key)
+        else:
+            print("Requested odp_key not found in odp_key data file")
+            return False
+        return data.keys()
 
     @property
     def body(self):